+++ domenetjenester/brevbestiller/pom.xml
@@ -71,12 +71,6 @@
         <dependency>
             <groupId>no.nav.foreldrepenger.melding</groupId>
             <artifactId>dokumentproduksjon-klient</artifactId>
-            <exclusions>
-                <exclusion>
-                <groupId>org.jvnet.jaxb2_commons</groupId>
-                <artifactId>jaxb2-basics-runtime</artifactId>
-                </exclusion>
-            </exclusions>
         </dependency>
         <dependency>
             <groupId>no.nav.foreldrepenger.kontrakter</groupId>
